+++ inc/vector_math/vector2f.hpp
@@ -37,7 +37,7 @@ namespace systems::leal::vector_math {
                 _mm_store_ps(result.data, _mm_add_ps(_mm_load_ps(data), _mm_load_ps(rhs.data)));
                 result.data[2] = result.data[3] = 0.0f;
                 return result;
-            #elif defined(__VECTOR_MATH_ARCH_ARM)
+            #elif defined(__VECTOR_MATH_ARCH_ARM) && (defined(__arm64__) || defined(__aarch64__))
                 Vector2f result;
                 vst1q_f32(result.data, vaddq_f32(vld1q_f32(data), vld1q_f32(rhs.data)));
                 result.data[2] = result.data[3] = 0.0f;
@@ -53,7 +53,7 @@ namespace systems::leal::vector_math {
                 _mm_store_ps(result.data, _mm_sub_ps(_mm_load_ps(data), _mm_load_ps(rhs.data)));
                 result.data[2] = result.data[3] = 0.0f;
                 return result;
-            #elif defined(__VECTOR_MATH_ARCH_ARM)
+            #elif defined(__VECTOR_MATH_ARCH_ARM) && (defined(__arm64__) || defined(__aarch64__))
                 Vector2f result;
                 vst1q_f32(result.data, vsubq_f32(vld1q_f32(data), vld1q_f32(rhs.data)));
                 result.data[2] = result.data[3] = 0.0f;
@@ -69,7 +69,7 @@ namespace systems::leal::vector_math {
                 _mm_store_ps(result.data, _mm_xor_ps(_mm_load_ps(data), _mm_set1_ps(-0.0f)));
                 result.data[2] = result.data[3] = 0.0f;
                 return result;
-            #elif defined(__VECTOR_MATH_ARCH_ARM)
+            #elif defined(__VECTOR_MATH_ARCH_ARM) && (defined(__arm64__) || defined(__aarch64__))
                 Vector2f result;
                 vst1q_f32(result.data, vnegq_f32(vld1q_f32(data)));
                 result.data[2] = result.data[3] = 0.0f;
@@ -85,7 +85,7 @@ namespace systems::leal::vector_math {
                 _mm_store_ps(result.data, _mm_mul_ps(_mm_load_ps(data), _mm_set1_ps(scalar)));
                 result.data[2] = result.data[3] = 0.0f;
                 return result;
-            #elif defined(__VECTOR_MATH_ARCH_ARM)
+            #elif defined(__VECTOR_MATH_ARCH_ARM) && (defined(__arm64__) || defined(__aarch64__))
                 Vector2f result;
                 vst1q_f32(result.data, vmulq_n_f32(vld1q_f32(data), scalar));
                 result.data[2] = result.data[3] = 0.0f;
@@ -103,7 +103,7 @@ namespace systems::leal::vector_math {
                 _mm_store_ps(result.data, _mm_div_ps(_mm_load_ps(data), _mm_set1_ps(scalar)));
                 result.data[2] = result.data[3] = 0.0f;
                 return result;
-            #elif defined(__VECTOR_MATH_ARCH_ARM)
+            #elif defined(__VECTOR_MATH_ARCH_ARM) && (defined(__arm64__) || defined(__aarch64__))
                 Vector2f result;
                 float inv = 1.0f / scalar;
                 vst1q_f32(result.data, vmulq_n_f32(vld1q_f32(data), inv));